Continue code clean-up
