+++ src/pysiral/auxdata/snow.py
@@ -49,7 +49,7 @@ from pyproj import Proj
 from xarray import open_dataset
 
 from pysiral.auxdata import AuxdataBaseClass, GridTrackInterpol
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import ErrorStatus
 from pysiral.core.iotools import ReadNC
 from pysiral.filter import idl_smooth
