+++ src/pysiral/core/output.py
@@ -18,7 +18,7 @@ from netCDF4 import Dataset, date2num
 from pysiral import psrlcfg
 from pysiral.core import DefaultLoggingClass
 from pysiral.core.config import get_yaml_config
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import ErrorStatus
 
 
 class OutputHandlerBase(DefaultLoggingClass):
