+++ src/pysiral/l2proc/procsteps.py
@@ -12,7 +12,7 @@ from loguru import logger
 from scipy.interpolate import interp1d
 
 from pysiral import get_cls
-from pysiral.core.class_template import DefaultLoggingClass
+from pysiral.core.legacy_classes import DefaultLoggingClass
 from pysiral.l1data import Level1bData
 from pysiral.l2data import Level2Data
