+++ src/pysiral/sentinel3/iotools.py
@@ -14,7 +14,7 @@ from parse import parse
 
 from pysiral.core import DefaultLoggingClass
 from pysiral.core.clocks import StopWatch
-from pysiral.core.errorhandler import ErrorStatus
+from pysiral.core.legacy_classes import ErrorStatus
 
 
 class Sentinel3FileList(DefaultLoggingClass):
